Refactor database to repository pattern
- Implement repository pattern with factory functions for users, sessions, auth tokens
- Update AppDatabase to use new appDb.users.create() interface style
- Update all 17 unit tests to use new repository interface
- All tests passing with improved code organization
