+++ lib/app-database.js
@@ -2,6 +2,9 @@ import { Database } from "bun:sqlite";
 import { readFileSync } from "node:fs";
 import { dirname, join } from "node:path";
 import { fileURLToPath } from "node:url";
+import { createAuthTokenRepository } from "./database/auth-tokens.js";
+import { createSessionRepository } from "./database/sessions.js";
+import { createUserRepository } from "./database/users.js";
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
@@ -9,12 +12,21 @@ export class AppDatabase {
   constructor(dbPath = process.env.DATABASE_PATH || "./app.db") {
     this.dbPath = dbPath;
     this.db = null;
+    // Repositories will be initialized after connection
+    this.users = null;
+    this.sessions = null;
+    this.authTokens = null;
   }
 
   async connect() {
     try {
       // Connect to existing database
       this.db = new Database(this.dbPath);
+
+      // Initialize repositories with shared db connection
+      this.users = createUserRepository(this.db);
+      this.sessions = createSessionRepository(this.db);
+      this.authTokens = createAuthTokenRepository(this.db);
 
       // Verify database exists and is properly initialized
       const needsInit = await this.needsInitialization();
@@ -36,6 +48,11 @@ export class AppDatabase {
     try {
       // Create database connection (creates file if it doesn't exist)
       this.db = new Database(this.dbPath);
+
+      // Initialize repositories with shared db connection
+      this.users = createUserRepository(this.db);
+      this.sessions = createSessionRepository(this.db);
+      this.authTokens = createAuthTokenRepository(this.db);
 
       // Check if database needs initialization
       const needsInit = await this.needsInitialization();
@@ -91,145 +108,9 @@ export class AppDatabase {
     if (this.db) {
       this.db.close();
       this.db = null;
-    }
-  }
-
-  // User management methods
-  async createUser(email) {
-    try {
-      const stmt = this.db.prepare(`
-        INSERT INTO users (email) 
-        VALUES (?) 
-        RETURNING id, email, created_at
-      `);
-
-      const user = stmt.get(email);
-      console.log(`👤 User created: ${email} (ID: ${user.id})`);
-      return user;
-    } catch (error) {
-      if (error.code === "SQLITE_CONSTRAINT_UNIQUE") {
-        throw new Error(`User with email ${email} already exists`);
-      }
-      console.error("❌ Failed to create user:", error);
-      throw error;
-    }
-  }
-
-  async getUserByEmail(email) {
-    try {
-      const stmt = this.db.prepare(`
-        SELECT id, email, created_at, last_login_at 
-        FROM users 
-        WHERE email = ?
-      `);
-
-      return stmt.get(email);
-    } catch (error) {
-      console.error("❌ Failed to get user by email:", error);
-      throw error;
-    }
-  }
-
-  async getUserById(id) {
-    try {
-      const stmt = this.db.prepare(`
-        SELECT id, email, created_at, last_login_at 
-        FROM users 
-        WHERE id = ?
-      `);
-
-      return stmt.get(id);
-    } catch (error) {
-      console.error("❌ Failed to get user by ID:", error);
-      throw error;
-    }
-  }
-
-  async updateLastLogin(userId) {
-    try {
-      const stmt = this.db.prepare(`
-        UPDATE users 
-        SET last_login_at = CURRENT_TIMESTAMP 
-        WHERE id = ?
-      `);
-
-      const result = stmt.run(userId);
-
-      if (result.changes === 0) {
-        throw new Error(`User with ID ${userId} not found`);
-      }
-
-      console.log(`👤 Updated last login for user ID: ${userId}`);
-      return result;
-    } catch (error) {
-      console.error("❌ Failed to update last login:", error);
-      throw error;
-    }
-  }
-
-  // Token management methods (for future phases)
-  // Note: token must be cryptographically secure (crypto.randomBytes(32).toString('hex'))
-  async createAuthToken(userId, token, expiresAt) {
-    try {
-      const stmt = this.db.prepare(`
-        INSERT INTO auth_tokens (user_id, token, expires_at) 
-        VALUES (?, ?, ?)
-        RETURNING id, token, expires_at
-      `);
-
-      return stmt.get(userId, token, expiresAt);
-    } catch (error) {
-      console.error("❌ Failed to create auth token:", error);
-      throw error;
-    }
-  }
-
-  async getAuthToken(token) {
-    try {
-      const stmt = this.db.prepare(`
-        SELECT t.id, t.token, t.user_id, t.expires_at, t.used_at, u.email
-        FROM auth_tokens t
-        JOIN users u ON t.user_id = u.id
-        WHERE t.token = ? AND t.used_at IS NULL
-      `);
-
-      return stmt.get(token);
-    } catch (error) {
-      console.error("❌ Failed to get auth token:", error);
-      throw error;
-    }
-  }
-
-  // Session management methods (for future phases)
-  // Note: sessionId must be cryptographically secure (crypto.randomBytes(32).toString('hex'))
-  async createSession(sessionId, userId, expiresAt) {
-    try {
-      const stmt = this.db.prepare(`
-        INSERT INTO sessions (id, user_id, expires_at) 
-        VALUES (?, ?, ?)
-        RETURNING id, user_id, expires_at
-      `);
-
-      return stmt.get(sessionId, userId, expiresAt);
-    } catch (error) {
-      console.error("❌ Failed to create session:", error);
-      throw error;
-    }
-  }
-
-  async getSession(sessionId) {
-    try {
-      const stmt = this.db.prepare(`
-        SELECT s.id, s.user_id, s.expires_at, u.email
-        FROM sessions s
-        JOIN users u ON s.user_id = u.id
-        WHERE s.id = ? AND datetime(s.expires_at) > datetime('now')
-      `);
-
-      return stmt.get(sessionId);
-    } catch (error) {
-      console.error("❌ Failed to get session:", error);
-      throw error;
+      this.users = null;
+      this.sessions = null;
+      this.authTokens = null;
     }
   }
